# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/camps/{campId}/tracks": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "전체 트랙 목록을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "캠프 ID",
                        "name": "campId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Track"
                            }
                        }
                    }
                }
            }
        },
...
        "/corners/{cornerId}/tracks": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 코너에 속한 트랙 목록을 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "코너별 트랙 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "코너 ID",
                        "name": "cornerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Track"
                            }
                        }
                    }
                }
            }
        },
...
        "/tracks": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 코너에 여러 트랙을 추가 생성한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 일괄 생성",
                "parameters": [
                    {
                        "description": "생성 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.CreateTracksRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_infrastructure_web.Track"
                            }
                        }
                    }
                }
            }
        },
        "/tracks/bulk-delete": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "선택한 트랙들을 일괄 삭제한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 일괄 삭제",
                "parameters": [
                    {
                        "description": "삭제할 트랙 ID 목록",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.BulkDeleteTracksRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "성공적으로 삭제됨"
                    }
                }
            }
        },
        "/tracks/export": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "인쇄를 위해 전체 트랙의 PIN 번호 등을 CSV로 다운로드한다.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 인증 정보 전체 내보내기",
                "responses": {
                    "200": {
                        "description": "CSV 데이터"
                    }
                }
            }
        },
        "/tracks/{id}": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "트랙 상세 정보(PIN 등)를 조회한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Track"
                        }
                    }
                }
            }
        },
        "/tracks/{id}/export": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 트랙의 PIN 번호를 PDF 형태로 다운로드한다.",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "단일 트랙 인증 정보 내보내기",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF 데이터"
                    }
                }
            }
        },
...
        "/tracks/{id}/regenerate-pin": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "특정 트랙의 PIN 번호를 새로 생성한다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "PIN 재발급",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.Track"
                        }
                    }
                }
            }
        },
        "/tracks/{id}/replace": {
            "put": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "기존 트랙을 삭제하고 지정한 대상 코너에 새 트랙을 생성하며 기존 진행자 세션의 마이그레이션 대상을 설정한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B. Resource Management (Admin)"
                ],
                "summary": "트랙 교체 (비상용)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "트랙 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "대상 코너",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ReplaceTrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ReplaceTrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_infrastructure_web.BulkDeleteTracksRequest": {
            "type": "object",
            "properties": {
                "trackIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_infrastructure_web.CreateTracksRequest": {
            "type": "object",
            "properties": {
                "campId": {
                    "type": "string"
                },
                "cornerId": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "internal_infrastructure_web.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 요청입니다."
                }
            }
        },
...
        "internal_infrastructure_web.ReplaceTrackRequest": {
            "type": "object",
            "properties": {
                "newCornerId": {
                    "type": "string"
                }
            }
        },
        "internal_infrastructure_web.ReplaceTrackResponse": {
            "type": "object",
            "properties": {
                "pin": {
                    "type": "string"
                },
                "track": {
                    "$ref": "#/definitions/internal_infrastructure_web.Track"
                }
            }
        },
...
        "internal_infrastructure_web.Track": {
            "type": "object",
            "properties": {
                "cornerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "currentVisit": {
                    "$ref": "#/definitions/internal_infrastructure_web.VisitSummary"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "operationalStatus": {
                    "type": "string",
                    "enum": [
                        "IDLE",
                        "BUSY"
                    ]
                },
                "pin": {
                    "type": "string",
                    "example": "482910"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "DELETED"
                    ]
                },
                "trackNo": {
                    "type": "integer"
                }
            }
        },
...
        "internal_infrastructure_web.VisitSummary": {
            "type": "object",
            "properties": {
                "cornerId": {
                    "type": "string",
                    "format": "uuid"
                },
                "deviationSeconds": {
                    "type": "integer"
                },
                "durationSeconds": {
                    "type": "integer"
                },
                "endedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "groupId": {
                    "type": "string",
                    "format": "uuid"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "inputMethod": {
                    "type": "string",
                    "enum": [
                        "QR_SCAN",
                        "MANUAL"
                    ]
                },
                "startedAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "IN_PROGRESS",
                        "COMPLETED"
                    ]
                },
                "trackId": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        }
    }
}